Telas de add e update, adicionada na tela de cliente, arquivos de usuario adicionado tambem, mas a tela esta com problema.

+++ frontend/src/services/clienteService.js
@@ -10,11 +10,15 @@ export const getAllClientes = async () => {
 };
 
 export const getClienteById = async (id_cliente) => {
+  if (!id_cliente) {
+    throw new Error('ID de cliente não definido.');
+  }
   try {
     const response = await api.get(`/cliente/view/${id_cliente}`);
     return response.data;
   } catch (error) {
-    throw new Error('Erro ao obter cliente.');
+    console.error('Erro ao buscar cliente:', error);
+    throw error;
   }
 };
 
@@ -28,6 +32,9 @@ export const createCliente = async (clienteData) => {
 };
 
 export const updateCliente = async (id_cliente, clienteData) => {
+  if (!id_cliente) {
+    throw new Error('id_cliente não definido.');
+  }
   try {
     const response = await api.put(`/cliente/update/${id_cliente}`, clienteData);
     return response.data;
@@ -37,6 +44,9 @@ export const updateCliente = async (id_cliente, clienteData) => {
 };
 
 export const deleteCliente = async (id_cliente) => {
+  if (!id_cliente) {
+    throw new Error('id_cliente não definido.');
+  }
   try {
     await api.delete(`/cliente/delete/${id_cliente}`);
   } catch (error) {
